+++ src/plang/interpreter/ExecuteBlock.java
@@ -2,9 +2,9 @@ package plang.interpreter;
 
 public final class ExecuteBlock {
     public final byte[] code;
-    public final int[] constantPool;
+    public final long[] constantPool;
 
-    public ExecuteBlock(byte[] code, int[] constantPool) {
+    public ExecuteBlock(byte[] code, long[] constantPool) {
         this.code = code;
         this.constantPool = constantPool;
     }
